Initial commit (old 2018 project)
Always fun to work with stuff from when I was a kid! I have some LEGO I want to make a map of Breath of the Wild from, and what's a better opportunity to turn this clunky Tkinter thing into something a bit more, uh, package-like with a CLI or whatever.

(To be fair when I was 18 I wanted to help my mum cross-stitch a photo of her dog.)